+++ src/Core/Filter/FieldFilter/EqualsFieldFilter.php
@@ -22,12 +22,13 @@ class EqualsFieldFilter extends AbstractFieldFilter
     /**
      * EqualsFieldFilter constructor.
      */
-    public function __construct(string $entity, array $condition = [])
+    public function __construct(string $entity, array $condition = [], string $alias = '')
     {
 
         $this->entity = $entity;
         $this->field = key($condition);
         $this->value = reset($condition);
+        $this->fieldAlias = $alias;
     }
 
 }