+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # Changelog
 
 - Implement cache with APCu, Array and Memcached/Couchebase adapters.
+- Implement Join and search by dependencies